Update README.md
Worked on this project of Country data clustering, built a Jupyter notebook with clusters and a PPT of the same. Clustered a list of 167 countries into 3 i.e. Developed, Developing and under-developed countries using socio-economic and health factors. I performed both K-means and Hierarchical clustering and listed the top-8 underdeveloped countries in the direst need of aid to the CEO of ABC International NGO